Renaming the Kiiroo 3rdParty class to KiirooGen2Vibe
Turns out the Pearl2 is the same family as the Fuse and
Blowbot. Now they're all in the same class.
